moved some methods from UserController to AttendeeController
