+++ src/main/controllers/AttendeeController.java
@@ -1,9 +1,11 @@
 package main.controllers;
 
 import main.entities.Attendee;
+import main.entities.Event;
 import main.usecases.ChatRoomManager;
 import main.usecases.MessageManager;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
@@ -12,8 +14,8 @@ import java.util.UUID;
  * checking and sending messages, and checking, registering, and dropping from events.
  *
  * @author Yi Tao Li
- * @version 1.1
- * @since 2020-11-09
+ * @version 1.2
+ * @since 2020-11-12
  */
 public class AttendeeController extends UserController{
 
@@ -64,24 +66,46 @@ public class AttendeeController extends UserController{
      *
      * @param user UUID of the other user who using is trying to remove
      */
-    public boolean removeUser(UUID user) {
+    public boolean removeContact(UUID user) {
         contactsManager.removeUser(this.loggedInUser, user);
         return true;
     }
 
     /**
-     * Sends a message to a contact and returns true.
-     *
-     * @param text String of the message that the user is trying to send
-     * @param chatRoom UUID of the ChatRoom that the user is trying to send their message to
-     * @param messageManager pre-defined MessageManager
-     * @param chatRoomManager pre-defined ChatRoomManager
+     * Returns a list of events in the conference.
      */
-    public boolean sendMessage(String text, UUID chatRoom, MessageManager messageManager,
-                               ChatRoomManager chatRoomManager) {
-        UUID message = messageManager.createMessage(text, this.loggedInUser);
-        chatRoomManager.addMessageToChatRoom(chatRoom, message);
-        return true;
+    public ArrayList<Event> checkEvents() {
+        return this.eventController.getAllEvents();
     }
 
+    /**
+     * Returns a list of events the user is registered in.
+     *
+     * @param eventController pre-defined EventController
+     */
+    public ArrayList<Event> checkRegisteredEvents(EventController eventController) {
+        return eventController.getUserEvents(this.loggedInUser);
+    }
+
+    /**
+     * Registers user in an event and returns true if they were not already registered.
+     *
+     * @param event UUID of the event that the user is trying to register for
+     */
+    public boolean registerForEvent(UUID event) {
+        // currently checkRegisteredEvents returns an ArrayList of Event and not UUID
+        if (!this.checkRegisteredEvents(this.eventController).contains(event)) {
+            return eventController.addUser(event, this.loggedInUser);
+        }
+        return false;
+    }
+
+    /**
+     * Removes an event from the user's list of registered events.
+     *
+     * @param event UUID of the event that the user is trying to remove
+     */
+    public boolean dropOutFromEvent(UUID event) {
+        return this.eventController.removeUser(event, this.loggedInUser);
+    }
 }
